Fix a bug of reset connection
